requirements.txt
python-telegram-bot
python-dotenv
youtube_dl
requests